% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glo_2ds.R
\name{daily_glo2ds_uri}
\alias{daily_glo2ds_uri}
\title{Retrieve the daily GLobal RTOFS glo2ds uri for the specifed date.  There is no test to
determine if the uri points to a valid resource.}
\usage{
daily_glo2ds_uri(date = Sys.Date() - 1, when = c("forecast",
  "nowcast")[1], what = c("prog", "diag")[1],
  root = "https://nomads.ncep.noaa.gov:9090/dods/rtofs")
}
\arguments{
\item{date}{character, Date or POSIXt, the date to retrieve, by default yesterday as Date class.
if provided as character it must be in YYYYmmdd format (no separator)}

\item{when}{character, either 'forecast' or 'nowcast'}

\item{what}{character, either 'prog' (prognosis) or 'diag' (diagnosis)}

\item{root}{character, the root uri}
}
\value{
ncdf4 class object or NULL
}
\description{
Retrieve the daily GLobal RTOFS glo2ds uri for the specifed date.  There is no test to
determine if the uri points to a valid resource.
}
