% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtofs-package.R
\docType{package}
\name{rtofs-package}
\alias{rtofs-package}
\alias{rtofs}
\title{rtofs-package}
\description{
Provides simplified access to OPeNDAP Real-time Ocean Forecast System
 (rtofs) data served by NCEP thredds servers.
}
