% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glo_2ds.R
\name{rtofs_daily_glo2ds}
\alias{rtofs_daily_glo2ds}
\title{Open a connection to an OPeNDAP resource}
\usage{
rtofs_daily_glo2ds(uri = daily_glo2ds_uri())
}
\arguments{
\item{uri}{character, uri for an opendap resource}
}
\value{
ncdf4 object or NULL
}
\description{
Open a connection to an OPeNDAP resource
}
