% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glo_2ds.R
\name{rtofs_glo2ds_get_raster}
\alias{rtofs_glo2ds_get_raster}
\title{Retrieve a Raster* for a user specified bounding box}
\usage{
rtofs_glo2ds_get_raster(NC = rtofs_daily_glo2ds(), param = "sst", ...)
}
\arguments{
\item{NC}{ncdf4 class object}

\item{param}{character the name of the parameter to extract}

\item{...}{further arguments for \code{\link{rtofs_glo2ds_nav}}}
}
\value{
raster of one or more layers or NULL
}
\description{
Retrieve a Raster* for a user specified bounding box
}
