% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glo_2ds.R
\name{rtofs_glo2ds_nav}
\alias{rtofs_glo2ds_nav}
\title{Compute the elements for navigating ncfd4 array extraction}
\usage{
rtofs_glo2ds_nav(
  NC,
  bb = c(-72, -63, 39, 46),
  level = 1,
  layers = seq_len(NC$dim$time$len),
  offset = 360
)
}
\arguments{
\item{NC}{ncdf4 class object}

\item{bb}{numeric, a 4 element vector specifing the bonding box in
[left, right, bottom, top] order.}

\item{level}{numeric vector or one or more contiguous level numbers.  Since
we are accessing surface data this is a bit of a red herring.  Best
to accept the default.}

\item{layers}{numeric vector or one or more contiguous layer numbers.}

\item{offset}{numeric the longitudinal offset need to translate between the
typical [-180, 180] to the rtofs degrees east system. By default it
is 360 which implies the input bb longitudes range form [-180, 180].
If the actual bb longitudes are [0,360] then set this to min(NC$dim$lon$vals)}
}
\value{
a list of navigation elements
\itemize{
\item{start a three element vector of start indices}
\item{count a three elemen count of the indices after start}
\item{lon a vector of the offset-translated longitude coordinates for the subset}
\item{lat a vector of the latitude coordinates for the subset}
\item{time a vector of POSIXct datetime vakues for each layer}
\item{res a two element of x and y resolution in rtofs native coordinates}
}
}
\description{
Compute the elements for navigating ncfd4 array extraction
}
