% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glo_2ds.R
\name{rtofs_root_uri}
\alias{rtofs_root_uri}
\title{Retrieve the root URI for the NOMADS NCEP rtofs resources}
\usage{
rtofs_root_uri(what = c("info", "opendap")[2])
}
\arguments{
\item{what}{character, one of "info" or "opendap" (default)}
}
\value{
a URI
}
\description{
Retrieve the root URI for the NOMADS NCEP rtofs resources
}
