% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glo_2ds.R
\name{rtofs_time}
\alias{rtofs_time}
\title{Retrieve a vector of date times for the layers in a NCDF4 object}
\usage{
rtofs_time(NC)
}
\arguments{
\item{NC}{ncdf4 class object}
}
\value{
a vector of Dates, POSIXct (one for each layer) or NULL
}
\description{
Retrieve a vector of date times for the layers in a NCDF4 object
}
